Move hook into separate helper
